features module renamed to feature_extraction
