+++ laserchicken/__init__.py
@@ -1,7 +1,7 @@
 from ._version import __version__
 
 from laserchicken.compute_neighbors import compute_neighborhoods
-from laserchicken.feature_extractor.features import compute_features, register_new_feature_extractor
+from laserchicken.feature_extractor.feature_extraction import compute_features, register_new_feature_extractor
 from laserchicken.io.load import load
 from laserchicken.io.export import export
 from laserchicken.build_volume import build_volume
