+++ laserchicken/feature_extractor/test_entropy_feature_extractor.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from laserchicken import compute_neighbors, keys, load, utils
 from laserchicken.feature_extractor.entropy_feature_extractor import EntropyFeatureExtractor
-from laserchicken.feature_extractor.features import compute_features
+from laserchicken.feature_extractor.feature_extraction import compute_features
 from laserchicken.test_tools import create_point_cloud
 from laserchicken.volume_specification import InfiniteCylinder
