+++ laserchicken/test_feature_extractor/test_feature_map.py
@@ -1,7 +1,7 @@
 """Test that the map from feature names to extractor classes is correct."""
 import unittest
 
-from laserchicken.feature_extractor import features
+from laserchicken.feature_extractor import feature_extraction
 from laserchicken.test_feature_extractor import Test1FeatureExtractor
 
 
@@ -18,22 +18,22 @@ class FeatureMapTests(unittest.TestCase):
                              'density_absolute_mean_z', 'density_absolute_mean_normalized_height', 'perc_15_z',
                              'perc_99_normalized_height', 'max_intensity', 'min_intensity', 'range_intensity', 'mean_intensity', 'std_intensity', 'coeff_var_intensity']
         for feature in expected_features:
-            self.assertIn(feature, features.FEATURES)
+            self.assertIn(feature, feature_extraction.FEATURES)
 
     def test_feature_map_contains_new_feature_after_registration(self):
         test_feature_extractor = Test1FeatureExtractor()
         expected_features = test_feature_extractor.provides()
 
-        features.register_new_feature_extractor(test_feature_extractor)
+        feature_extraction.register_new_feature_extractor(test_feature_extractor)
 
         for feature in expected_features:
-            self.assertIn(feature, features.FEATURES)
+            self.assertIn(feature, feature_extraction.FEATURES)
 
     def test_list_all_feature_names_contains_items(self):
-        names = features.list_feature_names()
+        names = feature_extraction.list_feature_names()
         self.assertGreater(len(names), 0)
 
     def test_list_all_feature_names_are_strings(self):
-        names = features.list_feature_names()
+        names = feature_extraction.list_feature_names()
         for name in names:
             self.assertEqual(type(name), str)
